backend/deploy-status/requirements.txt
psycopg2-binary>=2.9.0
paramiko>=3.0.0
